+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [33.3.6] 2023-06-01
+### Fixed
+- Fixed a bug that the 'system' value is missing in QuestionnaireResponse when
+  an item has an answerValueSet and it also has a questionnaire-item-contontrol
+  extension that is 'autocomplete' (search field).
+
 ## [33.3.5] 2023-05-17
 ### Fixed
 - Fixed a bug that FHIRPath expressions could run on a QuestionnaireResponse where